focusguard/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
aiosqlite==0.20.0
PyYAML==6.0.1
httpx==0.27.0
apscheduler==3.10.4
google-api-python-client==2.128.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
